// src/auth/dto/stripe-connect.dto.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';

export class UserStripeOnboardingLinkDto {
  @ApiProperty({ description: 'The onboarding URL generated by Stripe' })
  onboardingUrl: string;

  @ApiProperty({ description: 'Status message' })
  message: string;
}

export class UserStripeConnectStatusDto {
  @ApiProperty({
    description: 'Whether the user has a connected Stripe account',
  })
  isConnected: boolean;

  @ApiPropertyOptional({ description: 'The Stripe account ID if it exists' })
  stripeConnectId?: string;

  @ApiProperty({ description: 'Status message' })
  message: string;
}